Crawlers/requirements.txt
attrs==19.2.0 
scrapy==1.6.0
elasticsearch==6.2.0
scrapyd==1.2.0
Twisted==18.9.0
python-dateutil==2.8.0
